chatbot_system/requirements.txt
# Core Dependencies
google-generativeai>=0.3.0
pydantic>=2.5.0
pyyaml>=6.0.1

# Memory Backends
redis>=5.0.0
pymongo>=4.6.0
motor>=3.3.0  # Async MongoDB

# Vector Database (Optional)
pinecone-client>=3.0.0
chromadb>=0.4.0

# NLP & Text Processing
tiktoken>=0.5.0  # Token counting
langdetect>=1.0.9
textblob>=0.17.0  # Sentiment analysis

# API & Web
fastapi>=0.108.0
uvicorn>=0.25.0
httpx>=0.25.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0  # Retry logic
asyncio>=3.4.3
aioredis>=2.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
faker>=20.0.0

# Monitoring & Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Development
black>=23.12.0
ruff>=0.1.9
mypy>=1.8.0
pre-commit>=3.6.0

